requirements.txt
google.cloud.logging